Update README.md
This small project is a part of Harvard's CS50 course. It is a mini web app via which users can “buy” and “sell” stocks. The stocks (name and price) is fetched by foreign API (IEX) in a JSON format. This app enables the users to sign up, login and logout, to quote stock prices (realtime prices brought by https://iexcloud.io/) , to "buy" and "sell" stocks, to view history of actions, to view transactions and add "cash" to account. It kind of simulates those websites where users literally invest in stocks.